feat: change struktur folder atom > feature based
