+++ package.json
@@ -13,8 +13,6 @@
     "deploy": "gh-pages -d dist"
   },
   "dependencies": {
-    "@react-three/drei": "^10.0.8",
-    "@react-three/fiber": "^9.1.2",
     "@tailwindcss/vite": "^4.1.7",
     "email-validator": "^2.0.4",
     "feather-icons": "^4.29.2",
@@ -24,8 +22,7 @@
     "react-dom": "^19.1.0",
     "react-router": "^7.6.0",
     "react-typed": "^2.0.12",
-    "tailwindcss": "^4.1.7",
-    "three": "^0.176.0"
+    "tailwindcss": "^4.1.7"
   },
   "devDependencies": {
     "@eslint/js": "^9.25.0",
