--- src/assets/hero/text.css
@@ -1,11 +1,0 @@
-/* styles.css */
-.animated-text {
-  display: inline-block;
-  background: linear-gradient(270deg, #ff0080, #7928ca, #2afadf, #ff0080);
-  background-size: 800% 800%;
-  background-position: 0% 0%;
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-  color: transparent;
-  white-space: nowrap;
-}
