--- src/assets/index.js
@@ -1,4 +1,0 @@
-import hand from "./icons/hand.png";
-import aboutMe from "./icons/aboutMe.png";
-
-export { hand, aboutMe };
