--- src/components/Fragments/Elements/Paragraph.jsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-const Paragraph = (props) => {
-  const { className, children } = props;
-
-  return <p className={className}>{children}</p>;
-};
-
-export default Paragraph;
