--- src/components/Home.jsx
@@ -1,19 +1,0 @@
-import React from "react";
-
-// Layouts
-import MainLayout from "./layouts/MainLayout";
-
-// Organisms
-import Sections from "./organisms/Sections";
-
-function Home() {
-  return (
-    <>
-      <MainLayout>
-        <Sections />
-      </MainLayout>
-    </>
-  );
-}
-
-export default Home;
