+++ src/features/home/components/Jumbotron.jsx
@@ -3,13 +3,7 @@ import React, { useEffect, useRef } from "react";
 // gsap
 import { gsap } from "gsap";
 
-// Fragments
-import Heading from "../Fragments/Elements/Heading";
-import Paragraph from "../Fragments/Elements/Paragraph";
-
-const HeroSection = (props) => {
-  const { heroRef } = props;
-
+const Jumbotron = () => {
   // text element
   const elementCepi = useRef(null);
   const elementFront = useRef(null);
@@ -35,15 +29,15 @@ const HeroSection = (props) => {
 
   return (
     <section
-      ref={heroRef}
+      //   ref={heroRef}
       className="hero__section w-full h-screen flex items-center relative overflow-hidden"
     >
       <div className="wrapper__text relative lg:w-[50%] md:w-[60%] w-[80%] p-5 z-1">
-        <Heading className="sm:pt-3 pt-1 text-white sm:text-4xl text-4xl capitalize sm:w-[100%] w-[50%]">
+        <h1 className="sm:pt-3 pt-1 text-white sm:text-4xl text-4xl capitalize sm:w-[100%] w-[50%]">
           Hi! <span className="uppercase text-blue-500">i'm </span>
-        </Heading>
+        </h1>
 
-        <Heading
+        <h1
           ref={elementCepi}
           className="md:text-5xl text-4xl font-bold uppercase text-white text-shadow-[0_0_35px_rgb(255_255_255)]"
         >
@@ -60,20 +54,20 @@ const HeroSection = (props) => {
               {char}
             </span>
           ))}
-        </Heading>
+        </h1>
 
-        <Heading
+        <h1
           ref={elementFront}
           className="capitalize animated-text sm:pt-7 pt-0 sm:2xl text-2xl font-bold"
         >
           front end web
-        </Heading>
+        </h1>
 
-        <Paragraph className="font-[Lato,sans-serif] text-gray-300 md:text-lg sm:text-sm text-xs">
+        <p className="font-[Lato,sans-serif] text-gray-300 md:text-lg sm:text-sm text-xs">
           Hallo, saya cepi septiyana Front-End Web Developer Saya membangun
           antarmuka web yang cepat, responsif, dan user-friendly menggunakan
           teknologi modern seperti HTML, CSS, JavaScript, dan React.
-        </Paragraph>
+        </p>
       </div>
 
       {/* blob */}
@@ -84,4 +78,4 @@ const HeroSection = (props) => {
   );
 };
 
-export default HeroSection;
+export default Jumbotron;
