+++ src/features/home/data/icons.js
@@ -1,0 +1,36 @@
+import icon_Html from "../../../assets/tech/html-5.png";
+import icon_css from "../../../assets/tech/css-3.png";
+import icon_js from "../../../assets/tech/js.png";
+import icon_bootstrap from "../../../assets/tech/bootstrap.png";
+import icon_tailwind from "../../../assets/tech/tailwind.png";
+import icon_react from "../../../assets/tech/react.png";
+import icon_rtl from "../../../assets/tech/rtl.png";
+import icon_jest from "../../../assets/tech/jest.png";
+import icon_node from "../../../assets/tech/node.png";
+import icon_npm from "../../../assets/tech/npm.png";
+import icon_figma from "../../../assets/tech/figma.png";
+import icon_next from "../../../assets/tech/nextjs.svg";
+import icon_redux from "../../../assets/tech/redux.svg";
+import icon_reactRouter from "../../../assets/tech/react-router.svg";
+
+const technology = [
+  { name: "html", image: icon_Html },
+  { name: "css", image: icon_css },
+  { name: "javascript", image: icon_js },
+  { name: "react", image: icon_react },
+  { name: "node.js", image: icon_node },
+  { name: "next.js", image: icon_next },
+];
+
+const tools = [
+  { name: "tailwind", image: icon_tailwind },
+  { name: "bootstrap", image: icon_bootstrap },
+  { name: "reactTestingLibrary", image: icon_rtl },
+  { name: "jest", image: icon_jest },
+  { name: "npm", image: icon_npm },
+  { name: "figma", image: icon_figma },
+  { name: "redux", image: icon_redux },
+  { name: "react-router", image: icon_reactRouter },
+];
+
+export { technology, tools };
