+++ src/features/home/data/product.js
@@ -1,0 +1,37 @@
+// Image Project
+import img_Ecomerce from "../../../assets/ecommerce.png";
+import img_movie from "../../../assets/movies.png";
+import img_todo from "../../../assets/todo.png";
+
+const project = [
+  {
+    id: 1,
+    image: img_Ecomerce,
+    judul: "E-Commerce Dashboard",
+    desc: "Sebuah platform belanja online modern yang dirancang untuk memberikan pengalaman pengguna yang cepat, aman, dan responsif. Fitur utama mencakup katalog produk dinamis, keranjang belanja, sistem pembayaran terintegrasi, manajemen pengguna, dan dashboard admin",
+    tech: ["Vue.js", "Bootstrap"],
+    visit: "visit website",
+    link: "https://cepiseptiyana.github.io/Pangcuy/",
+  },
+  {
+    id: 2,
+    image: img_movie,
+    judul: "Landing Page Movies",
+    desc: "Platform pencarian dan eksplorasi film terkini. Temukan informasi lengkap, trailer, rating, dan ulasan dari berbagai genre film favoritmu.",
+    tech: ["React.js", "Tailwinds"],
+    visit: "visit website",
+    link: "https://cepiseptiyana.github.io/reactMv/",
+  },
+
+  {
+    id: 3,
+    image: img_todo,
+    judul: "Todo List",
+    desc: "Aplikasi ini dirancang untuk membantu pengguna mengatur aktivitas sehari-hari dengan antarmuka yang bersih dan responsif. Dibangun sebagai proyek portfolio untuk mendemonstrasikan kemampuan saya dalam pengembangan front-end modern menggunakan React.js",
+    tech: ["React.js", "Tailwinds"],
+    visit: "visit website",
+    link: "https://dicoding-react-dasar.vercel.app/",
+  },
+];
+
+export { project };
