+++ src/features/home/data/skill.teknis.js
@@ -1,0 +1,13 @@
+const skill_Teknis = [
+  "🔬 Atomic Design: Menerapkan prinsip desain komponen UI modular dan reusable berdasarkan struktur atom, molekul, organisme, template, dan halaman",
+  "📁 Feature-Based Folder Structure: Mengorganisir kode berdasarkan fitur/domain alih-alih tipe file, untuk meningkatkan skalabilitas, keterbacaan, dan kolaborasi tim pada proyek berskala besar.",
+  "🚀 CI/CD (Continuous Integration & Deployment): Mengotomatisasi proses build, test, dan deployment menggunakan GitHub Actions",
+  "📦 Component-Based Architecture: Membangun aplikasi berbasis komponen mandiri dan scalable, terutama dalam React/Vue.",
+  "⚙️ Version Control (Git): Kolaborasi tim menggunakan branching workflow, PR review, dan git rebase / squash dengan GitHub.",
+  "🎨 Design System Integration: Bekerja dengan Figma untuk mendefinisikan dan menyinkronkan UI guidelines antar tim dev/designer.",
+  "🧪 Automated Testing: Menulis unit & integration tests menggunakan Jest, React Testing Library,",
+  "🖌️ CSS & BEM Menulis stylesheet yang terstruktur, konsisten, dan mudah dipelihara menggunakan metodologi BEM (Block Element Modifier).",
+  "📊 Global State Integration (React-Redux): Menghubungkan state global ke komponen React menggunakan useSelector, useDispatch, dan context separation untuk performa optimal.",
+];
+
+export { skill_Teknis };
